Test fixed

+++ src/main/java/com/example/SimpleMath.java
@@ -13,7 +13,7 @@ public class SimpleMath {
         return a / b;
     }
 
-    public int multiply(int a, int b) {
+    public long multiply(int a, int b) {
         return a * b;
     }
 }
